Base version
/* 
PURPOSE: There is requirement for a product to implement GDPR compliance by moving out data from ODS/ODW/DWH. The challenge being dependency is not defined in any data model and there is no interface in the enterprise to delete dependent records. This is used as a template to write similar code in Oracle and other databases across organization. 
*/
Example call to the program 
Accepts the parameters in form of JSON for the source table details and also target archive details. Highlight of this API is ability to accept relationship to delete the records based on relationship defined in "tablerelationspath". It accepts the edges of all relation and builds a query to perform the deletion and movement into target schema. It also manages the transaction security while performing this operation.
*/

It works based on 
***JSON Properties***

OPERATION:  "DELETE" / "ARCHIVE",
SOURCESCHEMA: "schemaname"
ARCHTABLEDETAILS: {tablename, primary keys }
MASTERKEYDETAILS: 

*/
SAMPLE CALL
---------------

CALL archive_cdr ('
{
					"operation": "delete",
                    "archlevel":"study",
					"sourceschema":"gdpr",
                    "archtabledetails": {"tablename":"orders","keycolumn":["order","orderlinenumber"]}, 
					"masterkeydetails": {"keytable":"customers" , "keycolumn":"customernumber", "keyvalue":"181"}, 
					"tablerelationpath":[
										{ "righttable": "customers",
										  "rightkeys": ["customernumber"], 
										  "lefttable": "orders",
										  "leftkeys": ["customernumber"]
										}   
								]
							}
','{
								"schema":"gdpr",
								"tablename":"archive_orderdetails"
						}');